Open repository and edit link in new tab
When editing a page on GitHub (via the edit link), I often find myself referring back to the pages before/after the page I'm editing. For this reason, I added the `target="_blank"` attribute to the link a long time ago for Comprehensive Rust.

I suggest adding it to the template so everybody gets this behavior. I also made the link back to the repository open in a new tab for consistency (the print page is left as-is since it doesn't navigate away from the book).